`#wip` range join tests
`testOnly com.gitlab.leafty.test.misc.RangeJoinSpec` works

+++ src/test/scala/com/gitlab/leafty/test/misc/package.scala
@@ -5,7 +5,7 @@ import java.sql.Timestamp
 
 package object misc {
 
-  case class Trn(amount: Integer, time: Timestamp)
+  case class Trn(amount: BigDecimal, time: Timestamp)
 
   case class Range(start: Timestamp, end: Timestamp)
